pyproject.toml
[project]
name = "moex-portfolio"
version = "0.1.0"
description = "MOEX portfolio research and construction (CSV-auditable), using moexapi."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "moexapi>=1.0",
  "numpy>=1.26",
  "pandas>=2.2",
  "scipy>=1.12",
  "pyyaml>=6.0.1",
  "streamlit>=1.30",
  "matplotlib>=3.8",
]

[project.scripts]
moexpf = "moex_portfolio.cli:main"

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

